# optimized _modified_bessel_i and finished todo