did as the todo said, moved the filtering of the zero values to the top of the loop to prevent redundant operations, also used PyTorch's fused arithmetic operations to optimize the function further

https://colab.research.google.com/drive/1jNVehlZJ2AXn0XVqV5Ij9SMZ_XPMOfxB?usp=sharing
Here's the benchmarking+validation script

[Benchmark 1] DENSE tensor (size: 10000, 0% zeros)
--------------------------------------------------
Compute-then-Mask    : 10.2011 ms per run
Filter-then-Compute  : 3.0826 ms per run
🏆 Verdict: Filter-then-Compute is 3.31x faster on dense data.

[Benchmark 2] SPARSE tensor (size: 10000, 50% zeros)
--------------------------------------------------
Compute-then-Mask    : 22.1584 ms per run
Filter-then-Compute  : 1.8892 ms per run
🏆 Verdict: Filter-then-Compute is 11.73x faster on sparse data.